Rename dkenv package to lib
This avoids a conflict when building, as the binary is named dkenv.
